minimized usage of newline
