+++ app/src/main/java/com/example/cocktaildictionary/actions/Action.kt
@@ -1,0 +1,4 @@
+package com.example.cocktaildictionary.actions
+
+interface Action {
+}
