+++ app/src/main/java/com/example/cocktaildictionary/network/CocktailList.kt
@@ -1,0 +1,7 @@
+package com.example.cocktaildictionary.network
+import com.google.gson.annotations.SerializedName
+
+data class CocktailList(
+    @SerializedName("drinks") var Cocktails: List<Cocktail>
+) {
+}
