+++ app/src/main/java/com/example/cocktaildictionary/network/RetrofitClientInstance.kt
@@ -1,0 +1,24 @@
+package com.example.cocktaildictionary.network
+
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
+
+object RetrofitClientInstance {
+
+    private var retrofit: Retrofit? = null
+    private const val BASE_URL = "https://the-cocktail-db.p.rapidapi.com/"
+
+
+    val retrofitInstance: Retrofit?
+        get() {
+            if (retrofit == null) {
+                retrofit = Retrofit.Builder()
+                    .baseUrl(BASE_URL)
+                    .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .build()
+            }
+            return retrofit
+        }
+}
