+++ app/src/main/java/com/example/cocktaildictionary/state/LoadingViewState.kt
@@ -1,0 +1,15 @@
+package com.example.cocktaildictionary.state
+
+import com.example.cocktaildictionary.network.CocktailList
+
+
+/*
+*This is an implementation of [State] that describes the configuration of the loading screen at a given time
+ */
+data class LoadingViewState(
+    val showHeading : Boolean = true,
+    val showProgressCircle: Boolean = true,
+    val data: CocktailList? = null,
+    val error: Throwable? = null,
+    val isAppBeingRefreshed: Boolean? = false
+):State
