+++ app/src/main/java/com/example/cocktaildictionary/state/State.kt
@@ -1,0 +1,3 @@
+package com.example.cocktaildictionary.state
+
+interface State
