--- app/src/main/java/com/example/task1/actions/Action.kt
@@ -1,4 +1,0 @@
-package com.example.task1.actions
-
-interface Action {
-}
