--- app/src/main/java/com/example/task1/cache/Cache.kt
@@ -1,14 +1,0 @@
-package com.example.task1.cache
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.example.task1.network.Cocktail
-import com.example.task1.network.CocktailList
-
-@Entity(tableName = "Cache")
-data class Cache(
-
-    @PrimaryKey(autoGenerate = true) val id:Int?,
-    val time: String,
-    val latestData: String
-)
