--- app/src/main/java/com/example/task1/network/Cocktail.kt
@@ -1,13 +1,0 @@
-package com.example.task1.network
-
-import com.google.gson.annotations.SerializedName
-
-
-data class Cocktail(
-    @SerializedName("idDrink") val id: String,
-    @SerializedName("strDrinkThumb") val img : String,
-    @SerializedName("strDrink") val drinkName: String,
-    @SerializedName("strInstructions") val drinkInstruction:String
-)
-
-
