--- app/src/main/java/com/example/task1/network/CocktailList.kt
@@ -1,7 +1,0 @@
-package com.example.task1.network
-import com.google.gson.annotations.SerializedName
-
-data class CocktailList(
-    @SerializedName("drinks") var Cocktails: List<Cocktail>
-) {
-}
