--- app/src/main/java/com/example/task1/reducer/Reducer.kt
@@ -1,8 +1,0 @@
-package com.example.task1.reducer
-
-import com.example.task1.actions.Action
-import com.example.task1.state.State
-
-interface Reducer<S:State,A:Action> {
-    fun reduce(currentState: S, action: A): S
-}
