--- app/src/main/java/com/example/task1/state/State.kt
@@ -1,3 +1,0 @@
-package com.example.task1.state
-
-interface State
