--- app/src/main/java/com/example/task1/utils/Converters.kt
@@ -1,24 +1,0 @@
-package com.example.task1.utils
-
-import androidx.room.TypeConverter
-import com.example.task1.network.Cocktail
-import com.google.gson.Gson
-
-import com.google.gson.reflect.TypeToken
-import java.lang.reflect.Type
-
-
-class Converters {
-    @TypeConverter
-    fun fromString(value: String?): List<Cocktail> {
-        val listType: Type = object : TypeToken<List<Cocktail>>() {}.type
-        return Gson().fromJson(value, listType)
-    }
-
-    @TypeConverter
-    fun fromCocktailList(list: List<Cocktail>): String {
-        val gson = Gson()
-        var json: String = gson.toJson(list)
-        return json
-    }
-}
